Change to log and restart prompt
Title and link of the image are now added to the log *before* downloading the image. That means you can still go look at the image if you stopped the program or it took too long to download etc.

Default option for restarting is set to "y", so an empty string (simply pressing enter) doesn't crash the program anymore.